Create database.cpp and remove birb.cpp
